==> analyses/5community_climate_ecology/analyses/Step_0_data_transformations.Rmd <==
---
title: "Data transformations specific to community analyses"
author: "Joshua G. Smith"
date: "2022-08-23"
output: html_document
---

#required packages
```{r}
rm(list=ls())
require(dplyr)
```

**Part 1 - central coast processing for ordination**
#step 0: load data
#step 1: filter by region4=='central'
#step 2: drop species that were never observed (or only found in other regions).
#step 3: merge kelp and upc datasets (inverts and algae).
#step 4: clean up data for ordination.
#step 5: ordinate. 



#============================ Step 0 =========================================#

```{r}

dat_dir <- "/home/shares/ca-mpa/data/sync-data/monitoring/processed_data/community_climate_derived_data" 


CCFRP <- read.csv(file.path(dat_dir, "CCFRP_mpa_year.csv"))
kelp_swath <- read.csv(file.path(dat_dir, "kelp_swath_mpa_year.csv"))
kelp_upc <- read.csv(file.path(dat_dir, "kelp_upc_mpa_year.csv"))
kelp_combined <- read.csv(file.path(dat_dir, "kelp_swath_upc_mpa_year.csv"))
kelp_fish <- read.csv(file.path(dat_dir, "kelp_fish_mpa_year.csv"))
deep_reef <- read.csv(file.path(dat_dir, "deep_reef_mpa_year.csv"))
rocky <- read.csv(file.path(dat_dir, "rocky_mpa_year.csv"))
rocky_airtemp <- readRDS("/home/shares/ca-mpa/data/sync-data/environmental/processed/air_temp_intertidal_sites.rds")


```


#========================== Step 1 & 2 ========================================#

```{r}

###NOTE: input data are already filtered to central coast and should not 
#contain species that were never encountered. However, data drop thresholds
#should be included at this stage. 

CCFRP_cen <- CCFRP %>% filter(region4=='central') %>%
  select(where(~ any(. != 0)))%>%
  filter(year>=2007)
kelp_swath_cen <- kelp_swath %>% filter(region4=='central') %>%
  select(where(~ any(. != 0)))%>%
  filter(year>=2007)
kelp_upc_cen <- kelp_upc %>% filter(region4=='central') %>%
  select(where(~ any(. != 0)))%>%
  filter(year>=2007)
kelp_fish_cen <- kelp_fish %>% filter(region4=='central') %>%
  select(where(~ any(. != 0)))%>%
  filter(year>=2007)
deep_reef_cen <- deep_reef %>% filter(region4=='central') %>%
  select(where(~ any(. != 0)))%>%
  filter(year>=2007)
rocky_cen <- rocky %>% filter(region4=='central') %>%
  select(where(~ any(. != 0)))%>%
  filter(year>=2007)

```

#============================= Step 3 ========================================#

#Standardize and merge kelp swath and UPC data
```{r}

#standardize swath to max
kelp_swath_stan <- decostand(kelp_swath_cen[30:92], method="max",
                             margin=2, na.rm=TRUE)

nrow(kelp_swath_cen) #check length
nrow(kelp_swath_stan) #check length

kelp_swath_stan1 <- cbind(kelp_swath_cen[1:29],kelp_swath_stan)%>%
                    select(!c('dictyoneurum_californicum_reticulatum')) #drop this spp because UPC counts are more accurate




#standardize upc to max
kelp_upc_stan <- decostand(kelp_upc_cen[30:69], method="max",
                             margin=2, na.rm=TRUE)

nrow(kelp_upc_cen) #check length
nrow(kelp_upc_stan) #check length

kelp_upc_stan1 <- cbind(kelp_upc_cen[1:29],kelp_upc_stan) %>%
                  select(!c('macrocystis_pyrifera',
                       'stylaster_californicus',
                       'stephanocystis_osmundacea')) #remove UPC species that are recorded by swath




#Join stan dat

kelp_upc_stan2 <- kelp_upc_stan1 %>%
                  select(!(9:29))

kelp_stan <- left_join(kelp_swath_stan1, kelp_upc_stan2, by=c("year",
                                                              "region3", "region4",
                                                              "affiliated_mpa",
                                                              "mpa_defacto_class",
                                                              "MHW",
                                                              "mpa_defacto_designation"))%>%
                        select(!(c(group.x, group.y)))%>%
                        mutate(group = "kelp_invalg")%>%
                        select(group, everything())
               

#check for any unmatched data
kelp_check <-  anti_join(kelp_swath_stan1, kelp_upc_stan2, by=c("year",
                                                              "region3", "region4",
                                                              "affiliated_mpa",
                                                              "mpa_defacto_class",
                                                              "MHW",
                                                              "mpa_defacto_designation"))


#path_aurora <- "/home/shares/ca-mpa/data/sync-data/monitoring/processed_data/community_climate_derived_data"
#path_relative <- here::here("analyses","5community_climate_ecology","output")
#write.csv(kelp_stan, file.path(path_relative, "kelp_swath_upc_mpa_year.csv"), row.names = FALSE)
```


#============================= Step 4 ========================================#



```{r}
#CCFRP processing--------------------------------------------------------------
CCFRP_process <- CCFRP_cen %>%
  rowwise() %>%
  dplyr::mutate(sum = sum(across(30:ncol(.)), na.rm = T)) %>% #check for any sites with all zeros
  filter(!(sum==0))%>% #then remove any sites will all zeros
  dplyr::select(-sum)%>%
  mutate(desig_state = paste(mpa_designation,MHW))%>%
  dplyr::select(desig_state, everything())%>%
  arrange(year, desig_state) #remove rows containing only zeros


#define grouping vars
CCFRP_group_vars <- CCFRP_process %>%
  dplyr::select(1:9)

#define envr vars
CCFRP_envr_vars <- CCFRP_process %>%
    dplyr::select(c(SST_anom="sst_monthly_anom",
                    SST_abs = "sst_monthly_obs",
                  #CUTI="cuti_monthly_anom",
                  BEUTI_anom="beuti_monthly_anom",
                  BEUTI_abs = "beuti_monthly_obs",
                  MOCI="annual_MOCI",
                  BT_anom = "bottomT_monthly_anom",
                  BT_abs = "bottomT_monthly_obs"))
      

#define data for ordination
CCFRP_ord_data <- CCFRP_process %>%
  dplyr::select(31:ncol(.))

#standardize to max
CCFRP_rel <- decostand(CCFRP_ord_data, method = "hellinger")

#generate a BC dissim matrix
CCFRP_distmat <- vegdist(CCFRP_rel, method = "bray", na.rm=T) 


#kelp invalg processing--------------------------------------------------

kelp_invalg <- kelp_combined %>%
                  mutate(desig_state = paste(mpa_defacto_designation,MHW))%>%
                  dplyr::select(desig_state, everything())%>%
                  filter(mpa_defacto_designation=="smr" | mpa_defacto_designation=="ref") 

#define grouping vars
kelp_invalg_group_vars <- kelp_invalg%>%
                    dplyr::select(1:9)

#define envr vars
kelp_invalg_envr_vars <-  kelp_invalg %>%
  dplyr::select(c(SST_anom="sst_monthly_anom",
                    SST_abs = "sst_monthly_obs",
                  #CUTI="cuti_monthly_anom",
                  BEUTI_anom="beuti_monthly_anom",
                  BEUTI_abs = "beuti_monthly_obs",
                  MOCI="annual_MOCI",
                  BT_anom = "bottomT_monthly_anom",
                  BT_abs = "bottomT_monthly_obs"))


#define data for ordination
kelp_invalg_ord_data <- kelp_invalg%>%
              ungroup()%>%
              dplyr::select(31:ncol(.))


# NOTE decstand(method = "max") already performed in Step 2


#generate a BC dissim matrix
kelp_invalg_distmat <- 
    vegdist(kelp_invalg_ord_data, method = "bray", na.rm=T) #generates a BC dissim matrix




#kelp fish processing---------------------------------------------------------

kelp_fish <- kelp_fish_cen %>%
                  rowwise() %>%
                  dplyr::mutate(sum = sum(across(30:ncol(.)))) %>%
                  filter(!(sum==0))%>% #remove rows containing only zeros
                  dplyr::select(-sum) %>%
                  mutate(desig_state = paste(mpa_defacto_designation,MHW))%>%
                  dplyr::select(desig_state, everything())%>%
                  filter(mpa_defacto_designation=="smr" | mpa_defacto_designation=="ref") #drop sum columns and non-species categories  
         

#define grouping vars
kelp_fish_group_vars <- kelp_fish %>%
                    dplyr::select(1:9)

#define envr vars
kelp_fish_envr_vars <-  kelp_fish %>%
    dplyr::select(c(SST_anom="sst_monthly_anom",
                    SST_abs = "sst_monthly_obs",
                  #CUTI="cuti_monthly_anom",
                  BEUTI_anom="beuti_monthly_anom",
                  BEUTI_abs = "beuti_monthly_obs",
                  MOCI="annual_MOCI",
                  BT_anom = "bottomT_monthly_anom",
                  BT_abs = "bottomT_monthly_obs"))

#define data for ordination
kelp_fish_ord_data <- kelp_fish %>%
              ungroup() %>%
              dplyr::select(31:ncol(.))

#calculate relative abundance
kelp_fish_rel <- decostand(kelp_fish_ord_data, method = "hellinger")

#kelp_fish_rel <- kelp_fish_rel %>% slice_sample(n=2000) # testing if work on smaller

#generate a BC dissim matrix
kelp_fish_distmat <- 
    vegdist(kelp_fish_rel, method = "bray", na.rm=T) #generates a BC dissim matrix






#kelp swath processing---------------------------------------------------------

kelp_swath <- kelp_swath_cen %>%
                  rowwise() %>%
                  dplyr::mutate(sum = sum(across(30:ncol(.)))) %>%
                  filter(!(sum==0))%>% #remove rows containing only zeros
                  dplyr::select(-sum) %>%
                  mutate(desig_state = paste(mpa_defacto_designation,MHW))%>%
                  dplyr::select(desig_state, everything())%>%
                  filter(mpa_defacto_designation=="smr" | mpa_defacto_designation=="ref")
             

#define grouping vars
kelp_swath_group_vars <- kelp_swath %>%
                    dplyr::select(1:9)

#define envr vars
kelp_swath_envr_vars <-  kelp_swath %>%
   dplyr::select(c(SST_anom="sst_monthly_anom",
                    SST_abs = "sst_monthly_obs",
                  #CUTI="cuti_monthly_anom",
                  BEUTI_anom="beuti_monthly_anom",
                  BEUTI_abs = "beuti_monthly_obs",
                  MOCI="annual_MOCI",
                  BT_anom = "bottomT_monthly_anom",
                  BT_abs = "bottomT_monthly_obs"))

#define data for ordination
kelp_swath_ord_data <- kelp_swath %>%
              ungroup() %>%
              dplyr::select(31:ncol(.))

#calculate relative abundance
kelp_swath_rel <- decostand(kelp_swath_ord_data, method = "hellinger")


#generate a BC dissim matrix
kelp_swath_distmat <- 
    vegdist(kelp_swath_rel, method = "bray", na.rm=T) #generates a BC dissim matrix




#kelp upc processing---------------------------------------------------------

kelp_upc <- kelp_upc_cen %>%
                  rowwise() %>%
                  dplyr::mutate(sum = sum(across(30:ncol(.)))) %>%
                  filter(!(sum==0))%>% #remove rows containing only zeros
                  dplyr::select(-sum) %>%
                  mutate(desig_state = paste(mpa_defacto_designation,MHW))%>%
                  dplyr::select(desig_state, everything())%>%
                  filter(mpa_defacto_designation=="smr" | mpa_defacto_designation=="ref")
             

#define grouping vars
kelp_upc_group_vars <- kelp_upc %>%
                    dplyr::select(1:9)

#define envr vars
kelp_upc_envr_vars <-  kelp_upc %>%
    dplyr::select(c(SST_anom="sst_monthly_anom",
                    SST_abs = "sst_monthly_obs",
                  #CUTI="cuti_monthly_anom",
                  BEUTI_anom="beuti_monthly_anom",
                  BEUTI_abs = "beuti_monthly_obs",
                  MOCI="annual_MOCI",
                  BT_anom = "bottomT_monthly_anom",
                  BT_abs = "bottomT_monthly_obs"))

#define data for ordination
kelp_upc_ord_data <- kelp_upc %>%
              ungroup() %>%
              dplyr::select(31:ncol(.))

#calculate relative abundance
kelp_upc_rel <- decostand(kelp_upc_ord_data, method = "hellinger")


#generate a BC dissim matrix
kelp_upc_distmat <- 
    vegdist(kelp_upc_rel, method = "bray", na.rm=T) #generates a BC dissim matrix
 


#deep reef processing---------------------------------------------------------

deep_reef <- deep_reef_cen %>%
                  rowwise() %>%
                  dplyr::mutate(sum = sum(across(31:ncol(.)))) %>%
                  filter(!(sum==0))%>% #remove rows containing only zeros
                  dplyr::select(-c(
                                young_of_year_10_cm_sebastes_sp,
                                sum))%>%
                  mutate(desig_state = paste(mpa_defacto_designation,MHW))%>%
                  dplyr::select(desig_state, everything())%>%
                  filter(mpa_defacto_designation=="smr" | mpa_defacto_designation=="ref") 
        
#define grouping vars
deep_reef_group_vars <- deep_reef%>%
                    dplyr::select(1:10)

#define envr vars
deep_reef_envr_vars <-  deep_reef %>%
   dplyr::select(c(SST_anom="sst_monthly_anom",
                    SST_abs = "sst_monthly_obs",
                  #CUTI="cuti_monthly_anom",
                  BEUTI_anom="beuti_monthly_anom",
                  BEUTI_abs = "beuti_monthly_obs",
                  MOCI="annual_MOCI",
                  BT_anom = "bottomT_monthly_anom",
                  BT_abs = "bottomT_monthly_obs"))

#define data for ordination
deep_reef_ord_data <- deep_reef%>%
              ungroup()%>%
              dplyr::select(32:ncol(.))

#calculate relative abundance
deep_reef_rel <- decostand(deep_reef_ord_data, method="hellinger")

#generate a BC dissim mat
deep_reef_distmat <- 
    vegdist(deep_reef_rel, method = "bray", na.rm=T) #generates a BC dissim matrix



#Intertidal processing---------------------------------------------------------

rocky_Tair <- rocky_airtemp %>% mutate(marine_site_name = tolower(marine_site_name)) %>%
              ungroup() %>%
              select(!c(georegion, site_code))

rocky_cen$year <- as.numeric(as.character(rocky_cen$year))

rocky_counts <- #rocky_count_enr %>%
                rocky_cen %>%
                left_join(rocky_Tair, by= c("site"="marine_site_name","year")) %>%
                dplyr::select(year, group, region3, region4, site, mpa_designation, affiliated_mpa, AT_anom, AT_abs, everything()) %>%
                  mutate(MHW = ifelse(year>=2014 & year<=2016, "during",ifelse(year<2014, "before","after")))%>%
                  mutate(desig_state = paste(mpa_designation,MHW))%>%
                  dplyr::select(desig_state, MHW, everything())%>%
                  dplyr::filter(region4=='central')
                

rocky_counts <- rocky_counts %>% filter(mpa_designation == "smr" | mpa_designation == 'ref') %>% droplevels()

#define grouping vars
rocky_group_vars <- rocky_counts%>%
                    dplyr::select(1:9)

#define envr vars
rocky_envr_vars <-  rocky_counts %>%
  dplyr::select(c(SST_anom="sst_monthly_anom",
                    SST_abs = "sst_monthly_obs",
                  #CUTI="cuti_monthly_anom",
                  BEUTI_anom="beuti_monthly_anom",
                  BEUTI_abs = "beuti_monthly_obs",
                  MOCI="annual_MOCI",
                  AT_anom,
                  AT_abs))

#define data for ordination
rocky_ord_data <- rocky_counts %>%
              ungroup() %>%
              dplyr::select(28:ncol(.))

#counts are already relative, so apply sqrt transform to match "hellinger
rocky_rel <- rocky_ord_data %>%
             mutate(across(1:ncol(.),  function(x) sqrt(x)))

#generate a BC dissim matrix
rocky_distmat <- 
    vegan::vegdist(rocky_rel, method = "bray", na.rm=T) #generates a BC dissim matrix


###save group vars
#output_path <-  "/home/shares/ca-mpa/data/sync-data/monitoring/processed_data/community_climate_derived_data/reprocessed_data_20230615" #new
#output_path <-  "/home/shares/ca-mpa/data/sync-data/monitoring/processed_data/community_climate_derived_data/" #old

output_path <- here::here("analyses","5community_climate_ecology","output")


#save(CCFRP_group_vars,kelp_invalg_group_vars,kelp_fish_group_vars,
#     kelp_swath_group_vars, kelp_upc_group_vars,
#  deep_reef_group_vars,rocky_group_vars,
#    file = file.path(output_path, "group_vars.rda"))

#save community matrices
#save(CCFRP_ord_data,kelp_invalg_ord_data,kelp_fish_ord_data,
#    kelp_swath_ord_data, kelp_upc_ord_data,
#   deep_reef_ord_data,rocky_ord_data,
#   file = file.path(output_path, "comm_data.rda"))


#save envr vars 
#save(CCFRP_envr_vars,kelp_invalg_envr_vars,kelp_fish_envr_vars,
#     kelp_swath_envr_vars, kelp_upc_envr_vars,
#  deep_reef_envr_vars, rocky_envr_vars,
#   file = file.path(output_path, "envr_vars.rda"))

#save distance matrices
#save(CCFRP_distmat,kelp_invalg_distmat,kelp_fish_distmat,
 #   kelp_swath_distmat, kelp_upc_distmat,
  #   deep_reef_distmat,rocky_distmat,
   # file = file.path(output_path, "distance_matrices_BC.rda"))
```


#============================= Step 5 ========================================#

```{R}
# setting the seed 
set.seed(123)
num_cores=20

message("Processing CCFRP_ord")
CCFRP_ord <- metaMDS(CCFRP_distmat, distance="bray", trymax = 300, k=3, parallel = num_cores) #ordinates the dist matrix
CCFRP_en <- envfit(CCFRP_ord, CCFRP_envr_vars, permutations=999, na.rm=TRUE) #generate environmental vectors
#CCFRP_ord <- metaMDS(CCFRP_distmat, previous.best = CCFRP_ord, noscore=0.2, trymax=100)
message("Done processing CCFRP_ord")

message("Processing kelp_invalg_ord")
kelp_invalg_ord <- metaMDS(kelp_invalg_distmat, distance="bray", trymax = 300, k=3, parallel = num_cores) #ordinates the dist matrix
kelp_invalg_en <- envfit(kelp_invalg_ord, kelp_invalg_envr_vars, permutations=999, na.rm=TRUE) 
message("Done processing kelp_invalg_ord")

message("Processing kelp_fish_ord")
kelp_fish_ord <- metaMDS(kelp_fish_distmat, distance="bray", k=3,trymax = 300, parallel = num_cores)
#kelp_fish_ord <- metaMDS(kelp_fish_distmat, previous.best = kelp_fish_ord, noshare=0.2, trymax=300, parallel = num_cores)
kelp_fish_en <- envfit(kelp_fish_ord, kelp_fish_envr_vars, permutations=999, na.rm=TRUE) 
message("Done processing kelp_fish_ord")

message("Processing kelp_swath_ord")
kelp_swath_ord <- metaMDS(kelp_swath_distmat, distance="bray", k=3,trymax = 300, parallel = num_cores)
#kelp_fish_ord <- metaMDS(kelp_fish_distmat, previous.best = kelp_fish_ord, noshare=0.2, trymax=300, parallel = num_cores)
kelp_swath_en <- envfit(kelp_swath_ord, kelp_swath_envr_vars, permutations=999, na.rm=TRUE) 
message("Done processing kelp_swath_ord")

message("Processing kelp_upc_ord")
kelp_upc_ord <- metaMDS(kelp_upc_distmat, distance="bray", k=3,trymax = 300, parallel = num_cores)
#kelp_fish_ord <- metaMDS(kelp_fish_distmat, previous.best = kelp_fish_ord, noshare=0.2, trymax=300, parallel = num_cores)
kelp_upc_en <- envfit(kelp_upc_ord, kelp_upc_envr_vars, permutations=999, na.rm=TRUE) 
message("Done processing kelp_upc_ord")

message("Processing deep_reef_ord")
deep_reef_ord <- metaMDS(deep_reef_distmat, distance="bray", k=3,  trymax = 100, parallel = num_cores)
deep_reef_en <- envfit(deep_reef_ord, deep_reef_envr_vars, permutations=999, na.rm=TRUE) 
message("Done processing deep_reef_ord")

message("Processing rocky_ord")
rocky_ord <- vegan::metaMDS(rocky_distmat, distance="bray",k=3, trymax = 300, parallel = 20)
rocky_en <- vegan::envfit(rocky_ord, rocky_envr_vars, permutations=999, na.rm=TRUE) 
message("Done processing rocky_ord")


#export to .RData
save(CCFRP_ord, kelp_invalg_ord, kelp_fish_ord, deep_reef_ord, rocky_ord, file = file.path(output_path, "bray_nmds_scores.rda"))

save(CCFRP_en, kelp_invalg_en, kelp_fish_en, deep_reef_en, rocky_en, file = file.path(output_path, "env_fit_scores_with_absolutes.rda"))

```






**Part 2 - central coast processing for model based comparisons**
#step 0: load processed central coast data
#step 1: reshape to long format
#step 2: add species affinities and target status



#============================ Step 0 =========================================#

```{r}
rm(list=ls())

dat_dir <- "/home/shares/ca-mpa/data/sync-data/monitoring/processed_data/community_climate_derived_data" 


CCFRP <- read.csv(file.path(dat_dir, "CCFRP_mpa_year.csv"))
kelp_swath <- read.csv(file.path(dat_dir, "kelp_swath_mpa_year.csv"))
kelp_upc <- read.csv(file.path(dat_dir, "kelp_upc_mpa_year.csv"))
kelp_combined <- read.csv(file.path(dat_dir, "kelp_swath_upc_mpa_year.csv"))
kelp_fish <- read.csv(file.path(dat_dir, "kelp_fish_mpa_year.csv"))
deep_reef <- read.csv(file.path(dat_dir, "deep_reef_mpa_year.csv"))
rocky <- read.csv(file.path(dat_dir, "rocky_mpa_year.csv"))
rocky_airtemp <- readRDS("/home/shares/ca-mpa/data/sync-data/environmental/processed/air_temp_intertidal_sites.rds")

```


#============================ Step 1 =========================================#

```{r}
CCFRP_long <- CCFRP %>%
              pivot_longer(cols=30:ncol(.), names_to="species",
                           values_to = "counts")%>%
              filter(year >= 2007)

kelp_combined_long <- kelp_combined %>%
              pivot_longer(cols=30:ncol(.), names_to="species",
                           values_to = "counts")%>%
              filter(year >= 2007)

kelp_swath_long <- kelp_swath %>%
              pivot_longer(cols=30:ncol(.), names_to="species",
                           values_to = "counts")%>%
              filter(year >= 2007)

kelp_upc_long <- kelp_upc %>%
              pivot_longer(cols=30:ncol(.), names_to="species",
                           values_to = "counts")%>%
              filter(year >= 2007)

kelp_fish_long <- kelp_fish %>%
              pivot_longer(cols=30:ncol(.), names_to="species",
                           values_to = "counts")%>%
              filter(year >= 2007)

deep_reef_long <- deep_reef %>%
              pivot_longer(cols=31:ncol(.), names_to="species",
                           values_to = "counts")%>%
              filter(year >= 2007)

rocky_Tair <- rocky_airtemp %>% mutate(marine_site_name = tolower(marine_site_name)) %>%
              ungroup() %>%
              select(!c(georegion, site_code))
rocky_long <- rocky %>%
                #join air temp data
                left_join(rocky_Tair, by= c("site"="marine_site_name","year")) %>%
                dplyr::select(year, group, region3, region4, site, mpa_designation, affiliated_mpa, AT_anom, 
                              AT_abs, everything()) %>%
              #make longer
              pivot_longer(cols=26:ncol(.), names_to="species",
                           values_to = "counts")%>%
              filter(year >= 2007)
```


#============================ Step 1 =========================================#

#load affinity table
```{r}

affinity_dat <- read.csv("/home/shares/ca-mpa/data/sync-data/species_traits/processed/thermal_affinities_and_target_status.csv")

affinity_dat$common_name <- tolower(affinity_dat$common_name)
affinity_dat$genusspecies <- tolower(affinity_dat$genusspecies)

rocky_dat <- read.csv("/home/shares/ca-mpa/data/sync-data/species_traits/processed/rocky_thermal_affin.csv")

```

#join affinities

```{r}

affinity_select1 <- affinity_dat %>% 
                   select(common_name, thermal_affinity, targeted)%>%
                   distinct(common_name, .keep_all=TRUE)%>%
                   mutate(common_name = gsub(" ", "_", common_name))

affinity_select2 <- affinity_dat %>% 
                   select(scientific_name=genusspecies, thermal_affinity, targeted)%>%
                   distinct(scientific_name, .keep_all=TRUE)%>%
                   mutate(scientific_name = gsub(" ", "_", scientific_name))

################################################################################
#CCFRP 

#check unmatched species
CCFRP1 <- anti_join(CCFRP_long, affinity_select1, by=c("species"="common_name"))%>%
          distinct(species)

#fix unmatched names
CCFRP_long$species <- recode_factor(CCFRP_long$species, 
                                    "striped_seaperch"="striped_surfperch")
CCFRP_long$species <- recode_factor(CCFRP_long$species, 
                                    "pacific_bonito"="eastern_pacific_bonito")
CCFRP_long$species <- recode_factor(CCFRP_long$species, 
                                    "jack_smelt"="jacksmelt")
CCFRP_long$species <- recode_factor(CCFRP_long$species, 
                                    "mackerel_family"="pacific_mackerel,_greenback_mackerel")

#join dat
CCFRP_join1 <- left_join(CCFRP_long, affinity_select1, by=c("species"="common_name"))

################################################################################
#kelp fish

#check unmatched species
kelp_fish1 <- anti_join(kelp_fish_long, affinity_select2, by=c("species"="scientific_name"))%>%
          distinct(species)

#fix unmatched names
kelp_fish_long$species <- recode_factor(kelp_fish_long$species, 
                                    "bathymasteridae"="bathymasteridae_spp")
kelp_fish_long$species <- recode_factor(kelp_fish_long$species, 
                                    "citharichthys"="citharichthys_spp")
kelp_fish_long$species <- recode_factor(kelp_fish_long$species, 
                                    "platyrhinoidis_triseriata"="platyrhinoides_triseriata")
kelp_fish_long$species <- recode_factor(kelp_fish_long$species, 
                                    "sebastes_chrysomelas_carnatus"="sebastes_chrysomelas/carnatus")
kelp_fish_long$species <- recode_factor(kelp_fish_long$species, 
                                    "sebastes_dallii"="sebastes_dalli")
kelp_fish_long$species <- recode_factor(kelp_fish_long$species, 
                                    "sebastes_serranoides_flavidus"="sebastes_serranoides,flavidus")
kelp_fish_long$species <- recode_factor(kelp_fish_long$species, 
                                    "sebastes_serranoides_flavidus_melanops"="sebastes_serranoides,flavidus,melanops")


kelp_fish_join1 <- left_join(kelp_fish_long, affinity_select2, by=c("species"="scientific_name"))


################################################################################
#kelp inverts and algae
#check unmatched species
kelp_combined1 <- anti_join(kelp_combined_long, affinity_select2, by=c("species"="scientific_name"))%>%
          distinct(species)

#fix unmatched names
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "balanus_nubilus"="balanus_nubilis")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "crassadoma_gigantea"="crassedoma_giganteum")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "cryptolithodes_sitchensis"="cryptolithoides_sitchensis")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "haliotis_walallensis"="haliotis_wallalensis")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "loxorhynchus_crispatus_scyra_acutifrons"="loxorhynchus_scyra_spp")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "metridium"="metridium_spp")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "pachycerianthus_fimbriatus"="pachycerianthus_fimbratus")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "pisaster_ochraceus"="pisaster_ochraceous")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "urticina"="urticina_spp")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "bryozoa"="bryozoan")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "coralline_algae_crustose"="coralline_algae_-crustose")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "coralline_algae_erect_articulated"="coralline_algae_-erect/articulated")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "desmarestia"="desmarestia_spp")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "diopatra_chaetopterus_spp"="diopatra/chaetopterus_spp")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "tunicate_colonial_compund_social"="tunicate_-colonial,compund,social")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "dictyoneurum_californicum_reticulatum"="dictyoneurum_californicum/reticulatum")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "phyllospadix"="phyllospadix_spp")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
                                    "red_algae_branching_flat_blade"="red_algae_(branching_flat_blade)")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
              "thylacodes_squamigerus_petaloconchus_montereyensis"="thylacodes_squamigerus/petaloconchus_montereyensis")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
              "tunicate_solitary"="tunicate_-solitary")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
              "cucumaria"="cucumaria_spp")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
              "dictyotales"="dictyotales_spp")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
              "red_algae_leaf_like"="red_algae_(leaf-like)")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
              "red_algae_encrusting"="red_algae_-encrusting")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
              "red_algae_cylindrical_branches"="red_algae_(cylindrical_branches)")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
              "red_algae_lacy_branching"="red_algae_(lacy_branching)")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
              "tubeworm_mat"="tubeworm_mat.")
kelp_combined_long$species <- recode_factor(kelp_combined_long$species, 
              "mytilus_spp"="mussel")

#join
kelp_combined_join1 <- left_join(kelp_combined_long, affinity_select2, by=c("species"="scientific_name"))


################################################################################
#kelp upc
#check unmatched species
kelp_upc1 <- anti_join(kelp_upc_long, affinity_select2, by=c("species"="scientific_name"))%>%
          distinct(species)



#fix unmatched names
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "balanus_nubilus"="balanus_nubilis")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "crassadoma_gigantea"="crassedoma_giganteum")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "cryptolithodes_sitchensis"="cryptolithoides_sitchensis")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "haliotis_walallensis"="haliotis_wallalensis")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "loxorhynchus_crispatus_scyra_acutifrons"="loxorhynchus_scyra_spp")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "metridium"="metridium_spp")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "pachycerianthus_fimbriatus"="pachycerianthus_fimbratus")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "pisaster_ochraceus"="pisaster_ochraceous")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "urticina"="urticina_spp")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "bryozoa"="bryozoan")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "coralline_algae_crustose"="coralline_algae_-crustose")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "coralline_algae_erect_articulated"="coralline_algae_-erect/articulated")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "desmarestia"="desmarestia_spp")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "diopatra_chaetopterus_spp"="diopatra/chaetopterus_spp")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "tunicate_colonial_compund_social"="tunicate_-colonial,compund,social")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "dictyoneurum_californicum_reticulatum"="dictyoneurum_californicum/reticulatum")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "phyllospadix"="phyllospadix_spp")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
                                    "red_algae_branching_flat_blade"="red_algae_(branching_flat_blade)")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
              "thylacodes_squamigerus_petaloconchus_montereyensis"="thylacodes_squamigerus/petaloconchus_montereyensis")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
              "tunicate_solitary"="tunicate_-solitary")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
              "cucumaria"="cucumaria_spp")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
              "dictyotales"="dictyotales_spp")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
              "red_algae_leaf_like"="red_algae_(leaf-like)")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
              "red_algae_encrusting"="red_algae_-encrusting")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
              "red_algae_cylindrical_branches"="red_algae_(cylindrical_branches)")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
              "red_algae_lacy_branching"="red_algae_(lacy_branching)")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
              "tubeworm_mat"="tubeworm_mat.")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
              "mytilus_spp"="mussel")
kelp_upc_long$species <- recode_factor(kelp_upc_long$species, 
              "tunicate_colonial_compund_social"="tunicate_-colonial,compund,social")

#join
kelp_upc_join1 <- left_join(kelp_upc_long, affinity_select2, by=c("species"="scientific_name"))




################################################################################
#kelp swath
#check unmatched species
kelp_swath1 <- anti_join(kelp_swath_long, affinity_select2, by=c("species"="scientific_name"))%>%
          distinct(species)



#fix unmatched names
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "balanus_nubilus"="balanus_nubilis")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "crassadoma_gigantea"="crassedoma_giganteum")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "cryptolithodes_sitchensis"="cryptolithoides_sitchensis")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "haliotis_walallensis"="haliotis_wallalensis")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "loxorhynchus_crispatus_scyra_acutifrons"="loxorhynchus_scyra_spp")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "metridium"="metridium_spp")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "pachycerianthus_fimbriatus"="pachycerianthus_fimbratus")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "pisaster_ochraceus"="pisaster_ochraceous")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "urticina"="urticina_spp")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "bryozoa"="bryozoan")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "coralline_algae_crustose"="coralline_algae_-crustose")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "coralline_algae_erect_articulated"="coralline_algae_-erect/articulated")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "desmarestia"="desmarestia_spp")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "diopatra_chaetopterus_spp"="diopatra/chaetopterus_spp")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "tunicate_colonial_compund_social"="tunicate_-colonial,compund,social")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "dictyoneurum_californicum_reticulatum"="dictyoneurum_californicum/reticulatum")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "phyllospadix"="phyllospadix_spp")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
                                    "red_algae_branching_flat_blade"="red_algae_(branching_flat_blade)")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
              "thylacodes_squamigerus_petaloconchus_montereyensis"="thylacodes_squamigerus/petaloconchus_montereyensis")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
              "tunicate_solitary"="tunicate_-solitary")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
              "cucumaria"="cucumaria_spp")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
              "dictyotales"="dictyotales_spp")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
              "red_algae_leaf_like"="red_algae_(leaf-like)")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
              "red_algae_encrusting"="red_algae_-encrusting")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
              "red_algae_cylindrical_branches"="red_algae_(cylindrical_branches)")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
              "red_algae_lacy_branching"="red_algae_(lacy_branching)")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
              "tubeworm_mat"="tubeworm_mat.")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
              "mytilus_spp"="mussel")
kelp_swath_long$species <- recode_factor(kelp_swath_long$species, 
              "tunicate_colonial_compund_social"="tunicate_-colonial,compund,social")

#join
kelp_swath_join1 <- left_join(kelp_swath_long, affinity_select2, by=c("species"="scientific_name"))




################################################################################
#deep reef
#check unmatched species
deep_reef1 <- anti_join(deep_reef_long, affinity_select2, by=c("species"="scientific_name"))%>%
          distinct(species)

#fix unmatched names
deep_reef_long$species <- recode_factor(deep_reef_long$species, 
                                    "sebastes_dallii"="sebastes_dalli")
deep_reef_long$species <- recode_factor(deep_reef_long$species, 
                                    "sebastes_mystinus_or_diaconus"="sebastes_mystinus")

#join
deep_reef_join1 <- left_join(deep_reef_long, affinity_select2, by=c("species"="scientific_name"))



################################################################################
#rocky intertidal 

rocky_affin <- rocky_dat %>% dplyr::select(code, thermal_affinity)

#check unmatched species
rocky1 <- anti_join(rocky_long, rocky_affin, by=c("species" = "code"))%>%
          distinct(species)

#join
rocky_join1 <- left_join(rocky_long, rocky_affin, by=c("species"="code"))

```


#export data
```{r}
datalist <- list(CCFRP = CCFRP_join1, kelp_combined = kelp_combined_join1, 
                 kelp_swath = kelp_swath_join1, kelp_upc = kelp_upc_join1,
        deep_reef = deep_reef_join1, rocky = rocky_join1)

save(CCFRP_join1, kelp_combined_join1, 
                 kelp_swath_join1, kelp_upc_join1,
     kelp_fish_join1,
        deep_reef_join1, 
     rocky_join1,
     file= file.path(here::here("analyses","5community_climate_ecology","output","all_groups_mpa_level_means_long.rda")))

```
